+++ clients/vscode-hlasmplugin/src/code_actions/configurationFilesActions.ts
@@ -13,52 +13,121 @@
  */
 
 import * as vscode from 'vscode';
-import { ConfigurationNodeDetails, ConfigurationNodes, anyConfigurationNodeExists } from '../configurationNodes';
-import { proc_grps_file, pgm_conf_file } from '../constants';
+import { ConfigurationNodeDetails, ConfigurationNodes, ConfigurationNodesWithoutWorkspace, SettingsConfigurationNodeDetails, SettingsConfigurationNodeDetailsWithoutWorkspace, anyConfigurationNodeExists } from '../configurationNodes';
 
-function codeActionFactory(create: boolean, filename: string | undefined, args: any[] | undefined): vscode.CodeAction {
-    const filesDescription: string = filename ? filename + ' configuration file' : 'configuration files';
+function filenameFromUri(uri: vscode.Uri): string {
+    return uri.path.substring(uri.path.lastIndexOf('/') + 1);
+}
 
+function createAllConfigurationFiles(ws: vscode.Uri, relativeUri: string, pgroup: string): vscode.CodeAction {
     return {
-        title: (create ? 'Create ' : 'Modify ') + filesDescription,
+        title: 'Create configuration files',
         command: {
-            title: (create ? 'Create ' : 'Open ') + filesDescription,
-            command: create ? 'extension.hlasm-plugin.createCompleteConfig' : 'vscode.open',
-            arguments: args
+            title: 'Create configuration files',
+            command: 'extension.hlasm-plugin.createCompleteConfig',
+            arguments: [ws, relativeUri, pgroup],
         },
         kind: vscode.CodeActionKind.QuickFix
     };
 }
 
-function generateProcGrpsCodeAction(procGrps: ConfigurationNodeDetails, wsUri: vscode.Uri): vscode.CodeAction {
-    return procGrps.exists ? codeActionFactory(false, proc_grps_file, [procGrps.uri]) : codeActionFactory(true, proc_grps_file, [wsUri, null, '']);
+function createConfigurationFile(filename: string, ws: vscode.Uri, relativeUri: string | null, pgroup: string | null): vscode.CodeAction {
+    const filesDescription: string = filename + ' configuration file';
+    return {
+        title: 'Create ' + filesDescription,
+        command: {
+            title: 'Create ' + filesDescription,
+            command: 'extension.hlasm-plugin.createCompleteConfig',
+            arguments: [ws, relativeUri, pgroup],
+        },
+        kind: vscode.CodeActionKind.QuickFix
+    };
+}
+
+function modifyConfigurationFile(uri: vscode.Uri): vscode.CodeAction {
+    const filename = filenameFromUri(uri);
+    const filesDescription: string = filename + ' configuration file';
+
+    return {
+        title: 'Modify ' + filesDescription,
+        command: {
+            title: 'Open ' + filesDescription,
+            command: 'vscode.open',
+            arguments: [uri]
+        },
+        kind: vscode.CodeActionKind.QuickFix
+    };
+}
+
+function modifySettings(s: SettingsConfigurationNodeDetails | SettingsConfigurationNodeDetailsWithoutWorkspace): vscode.CodeAction {
+    const title = `Modify settings key '${s.key}'`;
+    if (s.scope === vscode.ConfigurationTarget.WorkspaceFolder)
+        return {
+            title,
+            command: {
+                title,
+                command: 'vscode.open',
+                arguments: [vscode.Uri.joinPath(s.ws, '.vscode/settings.json')],
+            },
+        };
+    else
+        return {
+            title,
+            command: {
+                title,
+                command: s.scope === vscode.ConfigurationTarget.Workspace ? 'workbench.action.openWorkspaceSettingsFile' : 'workbench.action.openSettingsJson',
+                arguments: [{
+                    revealSetting: {
+                        key: s.key,
+                        edit: true,
+                    }
+                }],
+            },
+        };
+}
+
+function generateProcGrpsCodeAction(procGrps: ConfigurationNodeDetails | SettingsConfigurationNodeDetails, wsUri: vscode.Uri): vscode.CodeAction {
+    if (procGrps.exists)
+        if ('uri' in procGrps)
+            return modifyConfigurationFile(procGrps.uri);
+        else
+            return modifySettings(procGrps);
+    else
+        return createConfigurationFile(filenameFromUri(procGrps.uri), wsUri, null, '');
 }
 
 function generatePgmConfCodeAction(configNodes: ConfigurationNodes, wsUri: vscode.Uri, documentRelativeUri: string): vscode.CodeAction {
-    if (configNodes.pgmConf.exists)
-        return codeActionFactory(false, pgm_conf_file, [configNodes.pgmConf.uri]);
+    if (configNodes.pgmConf.exists) {
+        if ('uri' in configNodes.pgmConf)
+            return modifyConfigurationFile(configNodes.pgmConf.uri);
+        else
+            return modifySettings(configNodes.pgmConf);
+    }
     else {
         if (configNodes.bridgeJson.exists || configNodes.ebgFolder.exists) {
             // TODO: could we trigger B4G sync?
         }
-        return codeActionFactory(true, pgm_conf_file, [wsUri, documentRelativeUri, null]);
+        return createConfigurationFile(filenameFromUri(configNodes.pgmConf.uri), wsUri, documentRelativeUri, null);
     }
 }
 
-export function generateConfigurationFilesCodeActions(suggestProcGrpsChange: boolean, suggestPgmConfChange: boolean, configNodes: ConfigurationNodes, wsUri: vscode.Uri, documentRelativeUri: string): vscode.CodeAction[] {
+export function generateConfigurationFilesCodeActions(suggestProcGrpsChange: boolean, suggestPgmConfChange: boolean, config: { nodes: ConfigurationNodes, ws: vscode.Uri, documentRelativeUri: string } | { nodes: ConfigurationNodesWithoutWorkspace }): vscode.CodeAction[] {
     if (!suggestProcGrpsChange && !suggestPgmConfChange)
         return [];
 
-    if (!anyConfigurationNodeExists(configNodes))
-        return [codeActionFactory(true, undefined, [wsUri, documentRelativeUri, 'GRP1'])];
+    if (!('ws' in config))
+        return [modifySettings(config.nodes.procGrps), modifySettings(config.nodes.pgmConf)];
+
+    if (!anyConfigurationNodeExists(config.nodes))
+        return [createAllConfigurationFiles(config.ws, config.documentRelativeUri, 'GRP1')];
 
     const result: vscode.CodeAction[] = [];
 
     if (suggestProcGrpsChange)
-        result.push(generateProcGrpsCodeAction(configNodes.procGrps, wsUri));
+        result.push(generateProcGrpsCodeAction(config.nodes.procGrps, config.ws));
 
     if (suggestPgmConfChange)
-        result.push(generatePgmConfCodeAction(configNodes, wsUri, documentRelativeUri));
+        result.push(generatePgmConfCodeAction(config.nodes, config.ws, config.documentRelativeUri));
 
     return result;
 }